agent/pyproject.toml
[project]
name = "lazylogger-agent"
version = "0.1.0"
description = "AI agent that integrates Obsidian vault with Telegram via LangChain and MCP"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]",
    "langchain-mcp-adapters",
    "langgraph",
    "langgraph-prebuilt",
    "python-telegram-bot>=22.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "aiohttp>=3.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
